+++ packages/components/src/index.ts
@@ -1,6 +1,9 @@
 // Assets
 export * from './assets'
 
+// Hooks
+export * from './hooks'
+
 // Atoms
 export { default as Badge } from './atoms/Badge'
 export type { BadgeProps } from './atoms/Badge'
@@ -16,6 +19,8 @@ export { default as Label } from './atoms/Label'
 export type { LabelProps } from './atoms/Label'
 export { default as Link } from './atoms/Link'
 export type { LinkProps, LinkElementType } from './atoms/Link'
+export { default as Loader } from './atoms/Loader'
+export type { LoaderProps } from './atoms/Loader'
 export { default as List } from './atoms/List'
 export type { ListProps } from './atoms/List'
 export { default as Overlay } from './atoms/Overlay'
@@ -26,6 +31,8 @@ export { default as Radio } from './atoms/Radio'
 export type { RadioProps } from './atoms/Radio'
 export { default as Select } from './atoms/Select'
 export type { SelectProps } from './atoms/Select'
+export { default as Slider } from './atoms/Slider'
+export type { SliderProps } from './atoms/Slider'
 export { default as SROnly } from './atoms/SROnly'
 
 // Molecules
@@ -72,10 +79,22 @@ export type {
   DropdownItemProps,
   DropdownMenuProps,
 } from './molecules/Dropdown'
+export {
+  default as Gift,
+  GiftContent,
+  GiftImage,
+} from './molecules/Gift'
+export type {
+  GiftProps,
+  GiftContentProps,
+  GiftImageProps,
+} from './molecules/Gift'
 export { default as InputField } from './molecules/InputField'
 export type { InputFieldProps } from './molecules/InputField'
 export { default as LinkButton } from './molecules/LinkButton'
 export type { LinkButtonProps } from './molecules/LinkButton'
+export { default as Modal, ModalHeader, ModalBody } from './molecules/Modal'
+export type { ModalProps, ModalHeaderProps } from './molecules/Modal'
 export { default as RadioField } from './molecules/RadioField'
 export type { RadioFieldProps } from './molecules/RadioField'
 export { default as RadioGroup, RadioOption } from './molecules/RadioGroup'
@@ -116,3 +135,6 @@ export type {
   HeroImageProps,
   HeroHeadingProps,
 } from './organisms/Hero'
+
+export { default as PriceRange } from './organisms/PriceRange'
+export type { PriceRangeProps } from './organisms/PriceRange'